+++ Meadowlark-js/tests/config/e2e/jest.config.js
@@ -5,8 +5,8 @@ module.exports = {
   displayName: 'E2E Tests',
   ...defaultConfig,
   testMatch: ['<rootDir>/**/e2e/**/*.(spec|test).[jt]s?(x)'],
-  globalSetup: './tests/e2e/setup/Setup.ts',
-  globalTeardown: './tests/e2e/setup/Teardown.ts',
+  globalSetup: './tests/e2e/setup/SetupTestEnvironment.ts',
+  globalTeardown: './tests/e2e/setup/TeardownTestEnvironment.ts',
   coverageThreshold: {
     global: {
       branches: 52,
